revived Cell.E; separate genome.go
