+++ multicell/indiv.go
@@ -8,17 +8,8 @@ import (
 	"math/rand"
 )
 
-type Genome struct { //Genome of an individual
-	E Spmat //Environment cue effect on epigenome
-	F Spmat //Regulatory effect of epigenetic markers on gene expression
-	G Spmat //Regulatory effect of gene expression on epigenetic markers
-	H Spmat //Contribution of gene expression on higher order complexes
-	J Spmat //Interaction between higher order complexes
-	P Spmat //Resulting expressed phenotype
-	//Z  Spmat //Gene expression of offspring
-}
-
 type Cell struct { //A 'cell' is characterized by its gene expression and phenotype
+	E        Vec     // Env + noise
 	F        Vec     // Epigenetic markers
 	G        Vec     // Gene expression
 	H        Vec     // Higher order complexes
@@ -52,239 +43,14 @@ type Indiv struct { //An individual as an unicellular organism
 	Dp0e1      float64 // ||p(e0) - e1||
 }
 
-func NewGenome() Genome { //Generate new genome matrix ensemble
-	E := NewSpmat(ngenes, nenv+ncells)
-	F := NewSpmat(ngenes, ngenes)
-	G := NewSpmat(ngenes, ngenes)
-	H := NewSpmat(ngenes, ngenes)
-	J := NewSpmat(ngenes, ngenes)
-	P := NewSpmat(ngenes, nenv+ncells)
-	//Z := NewSpmat(ngenes, ngenes)
-
-	genome := Genome{E, F, G, H, J, P}
-
-	return genome
-}
-
-func (G *Genome) Randomize() {
-
-	if withCue {
-		G.E.Randomize(CueResponseDensity, sdE)
-	}
-
-	if epig {
-		G.F.Randomize(GenomeDensity, sdF)
-	}
-	G.G.Randomize(GenomeDensity, sdG)
-
-	if hoc {
-		G.H.Randomize(GenomeDensity, sdH)
-		if hoi {
-			G.J.Randomize(GenomeDensity, sdJ)
-		}
-	}
-	G.P.Randomize(CueResponseDensity, sdP)
-}
-
-func (G *Genome) Clear() { //Sets all entries of genome to zero
-	for _, r := range G.E.Mat {
-		for j := range r { //range over keys
-			delete(r, j)
-		}
-	}
-	for _, r := range G.F.Mat {
-		for j := range r { //range over keys
-			delete(r, j)
-		}
-	}
-	for _, r := range G.G.Mat {
-		for j := range r { //range over keys
-			delete(r, j)
-		}
-	}
-	for _, r := range G.H.Mat {
-		for j := range r { //range over keys
-			delete(r, j)
-		}
-	}
-	for _, r := range G.J.Mat {
-		for j := range r { //range over keys
-			delete(r, j)
-		}
-	}
-	for _, r := range G.P.Mat {
-		for j := range r { //range over keys
-			delete(r, j)
-		}
-	}
-	/*
-		for _, r := range G.Z.Mat {
-			for j := range r { //range over keys
-				delete(r, j)
-			}
-		}
-	*/
-}
-
-func (parent *Genome) Copy() Genome { //creates copy of genome
-	e := parent.E.Copy()
-	f := parent.F.Copy()
-	g := parent.G.Copy()
-	hg := parent.H.Copy()
-	hh := parent.J.Copy()
-	p := parent.P.Copy()
-	//z := parent.Z.Copy()
-
-	//genome := Genome{e, f, g, hg, hh, p, z}
-
-	genome := Genome{e, f, g, hg, hh, p}
-
-	return genome
-}
-
-func DiffGenomes(Gout, G1, G0 *Genome) { //Elementwise difference between two genomes
-	if withCue {
-		Gout.E = DiffSpmat(&G1.E, &G0.E)
-	}
-	if epig {
-		Gout.F = DiffSpmat(&G1.F, &G0.F)
-	}
-
-	Gout.G = DiffSpmat(&G1.G, &G0.G)
-
-	if hoc {
-		Gout.H = DiffSpmat(&G1.H, &G0.H)
-		if hoi {
-			Gout.J = DiffSpmat(&G1.J, &G0.J)
-		}
-	}
-
-	Gout.P = DiffSpmat(&G1.P, &G0.P)
-	//Gout.Z = DiffSpmat(&G1.Z, &G0.Z)
-	//Remark: Ensure that Gout is initialized and empty before applying operation
-}
-
-func (G *Genome) NormalizeGenome() Genome {
-	lambda2 := 0.0
-	eG := G.Copy()
-
-	if withCue {
-		for _, m := range G.E.Mat {
-			for _, v := range m {
-				lambda2 += v * v
-			}
-		}
-	}
-
-	if epig {
-		for _, m := range G.F.Mat {
-			for _, v := range m {
-				lambda2 += v * v
-			}
-		}
-	}
-
-	for _, m := range G.G.Mat {
-		for _, v := range m {
-			lambda2 += v * v
-		}
-	}
-
-	if hoc {
-		for _, m := range G.H.Mat {
-			for _, v := range m {
-				lambda2 += v * v
-			}
-		}
-
-		if hoi {
-			for _, m := range G.J.Mat {
-				for _, v := range m {
-					lambda2 += v * v
-				}
-			}
-		}
-	}
-
-	for _, m := range G.P.Mat {
-		for _, v := range m {
-			lambda2 += v * v
-		}
-	}
-	/*
-		for _, m := range G.Z.Mat {
-			for _, v := range m {
-				lambda2 += v * v
-			}
-		}
-	*/
-
-	if lambda2 == 0 {
-		return eG //avoid division by zero
-	}
-
-	lambda := math.Sqrt(lambda2)
-
-	if withCue {
-		for i, m := range eG.E.Mat {
-			for j := range m {
-				eG.E.Mat[i][j] = eG.E.Mat[i][j] / lambda
-			}
-		}
-	}
-
-	if epig {
-		for i, m := range eG.F.Mat {
-			for j := range m {
-				eG.F.Mat[i][j] = eG.F.Mat[i][j] / lambda
-			}
-		}
-	}
-
-	for i, m := range eG.G.Mat {
-		for j := range m {
-			eG.G.Mat[i][j] = eG.G.Mat[i][j] / lambda
-		}
-	}
-
-	if hoc {
-		for i, m := range eG.H.Mat {
-			for j := range m {
-				eG.H.Mat[i][j] = eG.H.Mat[i][j] / lambda
-			}
-		}
-		if hoi {
-			for i, m := range eG.J.Mat {
-				for j := range m {
-					eG.J.Mat[i][j] = eG.J.Mat[i][j] / lambda
-				}
-			}
-		}
-	}
-
-	for i, m := range eG.P.Mat {
-		for j := range m {
-			eG.P.Mat[i][j] = eG.P.Mat[i][j] / lambda
-		}
-	}
-	/*
-		for i, m := range eG.Z.Mat {
-			for j := range m {
-				eG.Z.Mat[i][j] = eG.Z.Mat[i][j] / lambda
-			}
-		}
-	*/
-	return eG
-}
-
 func NewCell(id int) Cell { //Creates a new cell given id of cell.
+	e := NewVec(nenv + ncells)
 	f := NewVec(ngenes)
 	g := NewVec(ngenes)
 	h := NewVec(ngenes)
 	p := NewCue(nenv, id)
-	pv := Ones(nenv + ncells)
-	scaleVec(pv, pInitVar, pv)
-	cell := Cell{f, g, h, p, pv, 0.0, 0}
+	pv := NewVec(nenv + ncells)
+	cell := Cell{e, f, g, h, p, pv, 0.0, 0}
 
 	return cell
 }
@@ -293,6 +59,7 @@ func (cell *Cell) Copy() Cell {
 	id := GetId(cell.P) //Extract id part of cell phenotype
 	cell1 := NewCell(id)
 
+	cell1.E = CopyVec(cell.E)
 	cell1.F = CopyVec(cell.F)
 	cell1.G = CopyVec(cell.G)
 	cell1.H = CopyVec(cell.H)
@@ -389,41 +156,6 @@ func (indiv *Indiv) Mutate() { //Mutates portion of genome of an individual
 		indiv.Genome.P.mutateSpmat(CueResponseDensity, sdP)
 	}
 	return
-}
-
-func Crossover(dadg, momg *Genome) (Genome, Genome) { //Crossover
-	ng0 := dadg.Copy()
-	ng1 := momg.Copy()
-	//nz0 := dadz
-	//nz1 := momz
-
-	for i := 0; i < ngenes; i++ {
-		r := rand.Float64()
-		if r < 0.5 {
-			e := ng0.E.Mat[i]
-			f := ng0.F.Mat[i]
-			g := ng0.G.Mat[i]
-			hg := ng0.H.Mat[i]
-			hh := ng0.J.Mat[i]
-			p := ng0.P.Mat[i]
-
-			ng0.E.Mat[i] = ng1.E.Mat[i]
-			ng0.F.Mat[i] = ng1.F.Mat[i]
-			ng0.G.Mat[i] = ng1.G.Mat[i]
-			ng0.H.Mat[i] = ng1.H.Mat[i]
-			ng0.J.Mat[i] = ng1.J.Mat[i]
-			ng0.P.Mat[i] = ng1.P.Mat[i]
-
-			ng1.E.Mat[i] = e
-			ng1.F.Mat[i] = f
-			ng1.G.Mat[i] = g
-			ng1.H.Mat[i] = hg
-			ng1.J.Mat[i] = hh
-			ng1.P.Mat[i] = p
-		}
-	}
-
-	return ng0, ng1
 }
 
 func Mate(dad, mom *Indiv) (Indiv, Indiv) { //Generates offspring
@@ -526,14 +258,14 @@ func (cell *Cell) DevCell(G Genome, env Cue) Cell { //Develops a cell given cue
 	g1 := NewVec(ngenes)
 	h1 := NewVec(ngenes)
 
-	renv := AddNoise2Cue(env, devNoise)
+	AddNoise2Cue(cell.E, env, devNoise)
 
 	for nstep := 1; nstep <= maxDevStep; nstep++ {
 		multMatVec(vf, G.G, g0)
 		if withCue { //Model with or without cues
 			if pheno_feedback { //If feedback is allowed
 
-				diffVecs(e_p, renv, cell.P)
+				diffVecs(e_p, cell.E, cell.P)
 
 				// Kalman gain
 				//pscale := cell.getPscale()
@@ -541,7 +273,7 @@ func (cell *Cell) DevCell(G Genome, env Cue) Cell { //Develops a cell given cue
 
 				multMatVec(ve, G.E, e_p)
 			} else {
-				multMatVec(ve, G.E, renv)
+				multMatVec(ve, G.E, cell.E)
 			}
 			addVecs(f1, vf, ve)
 		} else {
